Refactored the test suite to eliminate redundancy and improve maintainability.
Key Changes:

1. Created ct/test_base.py (154 lines) containing:
	- BaseCompileToolsTestCase - common setUp/tearDown logic
	- create_magic_parser() - factory for magic flag parsers
	- create_header_deps_parser() - factory for header dependency parsers
	- compare_direct_cpp_magic() - utility for comparing direct vs cpp magic parsers
	- compare_direct_cpp_headers() - utility for comparing direct vs cpp header parsers
	- _verify_one_exe_per_main() - common executable verification
2. Refactored 4 test files to use the new base utilities:
	- test_magicflags.py: 252→141 lines (-44%)
	- test_headerdeps.py: 657→629 lines (-4%)
	- test_magicpkgconfig.py: 207→179 lines (-14%)
	- test_magicinclude.py: 116→90 lines (-22%)
