+++ ct/test_magicinclude.py
@@ -6,33 +6,14 @@ import configargparse
 import ct.unittesthelper as uth
 import ct.utils
 import ct.cake
+import ct.test_base as tb
 
 # Although this is virtually identical to the test_cake.py, we can't merge
 # the tests due to memoized results.
 
 
-class TestMagicInclude(unittest.TestCase):
-    def setUp(self):
-        try:
-            if self._tmpdir is not None:
-                shutil.rmtree(self._tmpdir, ignore_errors=True)
-        except AttributeError:
-            pass
-        self._tmpdir = tempfile.mkdtemp()
+class TestMagicInclude(tb.BaseCompileToolsTestCase):
 
-    def _verify_one_exe_per_main(self, relativepaths):
-        actual_exes = set()
-        for root, dirs, files in os.walk(self._tmpdir):
-            for ff in files:
-                if ct.utils.isexecutable(os.path.join(root, ff)):
-                    actual_exes.add(ff)
-
-        expected_exes = {
-            os.path.splitext(os.path.split(filename)[1])[0]
-            for filename in relativepaths
-            if ct.utils.issource(filename)
-        }
-        self.assertSetEqual(expected_exes, actual_exes)
 
     def test_magicinclude(self):
         # This test is to ensure that the //#INCLUDE magic flag
@@ -66,9 +47,7 @@ class TestMagicInclude(unittest.TestCase):
         relativepaths = ["magicinclude/main.cpp"]
         self._verify_one_exe_per_main(relativepaths)
 
-        # Cleanup
         os.chdir(origdir)
-        shutil.rmtree(self._tmpdir, ignore_errors=True)
 
     def test_magicinclude_append(self):
         # This test is to ensure that the //#INCLUDE magic flag
@@ -102,13 +81,9 @@ class TestMagicInclude(unittest.TestCase):
         relativepaths = ["magicinclude/main.cpp"]
         self._verify_one_exe_per_main(relativepaths)
 
-        # Cleanup
         os.chdir(origdir)
-        shutil.rmtree(self._tmpdir, ignore_errors=True)
 
 
-    def tearDown(self):
-        uth.reset()
 
 
 if __name__ == "__main__":
